requirements.txt
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pandas==2.2.1
scikit-learn==1.4.2
supabase==2.15.0
firebase-admin==6.4.0
numpy==1.26.4
scipy==1.12.0 